--- a/go.mod
+++ b/go.mod
@@ -2,9 +2,12 @@ module github.com/ipfs/go-log
 
 require (
 	github.com/gogo/protobuf v1.2.1
-	github.com/mattn/go-colorable v0.1.1
 	github.com/opentracing/opentracing-go v1.0.2
+	github.com/pkg/errors v0.8.1 // indirect
 	github.com/stretchr/testify v1.3.0 // indirect
 	github.com/whyrusleeping/go-logging v0.0.0-20170515211332-0457bb6b88fc
-	golang.org/x/net v0.0.0-20190227160552-c95aed5357e7 // indirect
+	go.uber.org/atomic v1.4.0 // indirect
+	go.uber.org/multierr v1.1.0 // indirect
+	go.uber.org/zap v1.10.0
+	golang.org/x/net v0.0.0-20190620200207-3b0461eec859 // indirect
 )