--- a/levels.go
+++ b/levels.go
@@ -2,6 +2,8 @@ package log
 
 import "go.uber.org/zap/zapcore"
 
+// LogLevel represents a log severity level. Use the package variables as an
+// enum.
 type LogLevel zapcore.Level
 
 var (
@@ -14,6 +16,13 @@ var (
 	LevelFatal  = LogLevel(zapcore.FatalLevel)
 )
 
+// LevelFromString parses a string-based level and returns the corresponding
+// LogLevel.
+//
+// Supported strings are: DEBUG, INFO, WARN, ERROR, DPANIC, PANIC, FATAL, and
+// their lower-case forms.
+//
+// The returned LogLevel must be discarded if error is not nil.
 func LevelFromString(level string) (LogLevel, error) {
 	lvl := zapcore.InfoLevel // zero value
 	err := lvl.Set(level)