--- a/oldlog.go
+++ b/oldlog.go
@@ -19,18 +19,6 @@ func init() {
 	SetupLogging()
 }
 
-var ansiGray = "\033[0;37m"
-var ansiBlue = "\033[0;34m"
-
-// LogFormats defines formats for logging (i.e. "color")
-var LogFormats = map[string]string{
-	"nocolor": "%{time:2006-01-02 15:04:05.000000} %{level} %{module} %{shortfile}: %{message}",
-	"color": ansiGray + "%{time:15:04:05.000} %{color}%{level:5.5s} " + ansiBlue +
-		"%{module:10.10s}: %{color:reset}%{message} " + ansiGray + "%{shortfile}%{color:reset}",
-}
-
-var defaultLogFormat = "color"
-
 // Logging environment variables
 const (
 	// TODO these env names should be more general, IPFS is not the only project to