… go.sum
… writer/option.go
… writer/polite_json_formatter.go