Builds and works with go-ipfs, exposes zap functionality.