@Kubuxu 2019-07-08
Updated and rebased on master