@lanzafame 2019-08-28
@Kubuxu @frrist with all the projects that use `go-log` now moved over to using go modules, can we tag this as a new major release and move on?