@hsanjuan 2019-08-28
I have customized LogFormats because I could not read things in blue. This will also break that. But ok..