@raulk 2019-08-29
## Quick alignment comment

@whyrusleeping: IIRC in the "exposing zap vs. abstraction with a zap backend" discussion, you advocated for keeping an abstraction layer as a way to substitute the backend in alternative runtime environments, like WASM. Zap is pretty modular and we should be able to do that replacement _within_ Zap itself. Are you OK going in this direction?

@albrow: since 0x is already running go-libp2p on WASM, are you happy with this change? It should bring important optimisations.

@Kubuxu: do you have bandwidth to introduce a few go benchmarks to quantify the gains under different conditions (console, file, structured logging, different levels, etc.)?

---

Looks like the tests provoke a deadlock and a few data races: https://travis-ci.com/ipfs/go-log/builds/125061936