@Kubuxu 2019-08-29
Quick benchmarks:
```
benchmark                      old ns/op     new ns/op     delta
BenchmarkSimpleInfo-4          5598          4015          -28.28%
BenchmarkFormatInfo-4          5819          4523          -22.27%
BenchmarkFormatInfoMulti-4     4482          2633          -41.25%

benchmark                      old allocs     new allocs     delta
BenchmarkSimpleInfo-4          27             10             -62.96%
BenchmarkFormatInfo-4          27             11             -59.26%
BenchmarkFormatInfoMulti-4     28             12             -57.14%

benchmark                      old bytes     new bytes     delta
BenchmarkSimpleInfo-4          928           368           -60.34%
BenchmarkFormatInfo-4          1488          569           -61.76%
BenchmarkFormatInfoMulti-4     1464          562           -61.61%
```