@Kubuxu 2019-08-29
> Looks like the tests provoke a deadlock and a few data races:

Those were there, it is a race inside a test (not code) and race detector causes the lockup.