@Kubuxu 2019-08-29
I probably won't have time to benchmark it more (at least right now).